refactor: 이름 변경 ai -> recommendation
