+++ batch/src/main/java/store/_0982/batch/domain/recommendation/PersonalVectorRepository.java
@@ -1,4 +1,4 @@
-package store._0982.batch.domain.ai;
+package store._0982.batch.domain.recommendation;
 
 public interface PersonalVectorRepository {
     void saveAll(Iterable<? extends PersonalVector> vectors);
