+++ batch/src/main/java/store/_0982/batch/infrastructure/recommendation/PersonalVectorJpaRepository.java
@@ -1,7 +1,7 @@
-package store._0982.batch.infrastructure.ai;
+package store._0982.batch.infrastructure.recommendation;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import store._0982.batch.domain.ai.PersonalVector;
+import store._0982.batch.domain.recommendation.PersonalVector;
 
 import java.util.UUID;
