+++ recommendation/src/main/java/store/_0982/recommendation/domain/PersonalVector.java
@@ -15,7 +15,7 @@ import java.time.OffsetDateTime;
 import java.util.UUID;
 
 @Entity
-@Table(name = "personal_vector", schema = "ai_schema")
+@Table(name = "personal_vector", schema = "recommendation_schema")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 public class PersonalVector {
